Remove cooldown reduction. Replace with more context. (#476)
tl;dr

* Staked HNT continues to become veHNT with 6 month lockup
* Staked HNT in Cooldown becomes locked HNT with vesting matching remaining cooldown period
* No change to any unstaked HNT.

Reasoning: We've received an overwhelming amount of feedback from validators / stakers that the reduction in cooldown, although preferable to some parties is the ecosystem, is an overall negative change in this HIP. Specifically, it introduces a major economic change to the network inside of a technical upgrade. Instead of introducing the cooldown, we've consulted with experts in the Solana ecosystem to implement the migration of stakes in cooldown to vesting tokens. This will enable anyone to unstake, if they choose to, at anytime and still be able to complete their cooldown after the L1 migration and accounts are imported. The cooldown period was agreed to by all validators when they initially staked.